mcp-server/pyproject.toml
[project]
name = "boltastic-mcp-server"
version = "0.1.0"
description = "Boltastic MCP server: UI analysis tool with screenshots and Gemini"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.13.0",
    "google-generativeai>=0.8.0",
    "playwright>=1.49.0",
]

[project.optional-dependencies]
dev = []
